// src/strava/Api.ts
export enum SportType {
  AlpineSki = "AlpineSki",
  BackcountrySki = "BackcountrySki",
  Badminton = "Badminton",
  Canoeing = "Canoeing",
  Crossfit = "Crossfit",
  EBikeRide = "EBikeRide",
  Elliptical = "Elliptical",
  EMountainBikeRide = "EMountainBikeRide",
  Golf = "Golf",
  GravelRide = "GravelRide",
  Handcycle = "Handcycle",
  HighIntensityIntervalTraining = "HighIntensityIntervalTraining",
  Hike = "Hike",
  IceSkate = "IceSkate",
  InlineSkate = "InlineSkate",
  Kayaking = "Kayaking",
  Kitesurf = "Kitesurf",
  MountainBikeRide = "MountainBikeRide",
  NordicSki = "NordicSki",
  Pickleball = "Pickleball",
  Pilates = "Pilates",
  Racquetball = "Racquetball",
  Ride = "Ride",
  RockClimbing = "RockClimbing",
  RollerSki = "RollerSki",
  Rowing = "Rowing",
  Run = "Run",
  Sail = "Sail",
  Skateboard = "Skateboard",
  Snowboard = "Snowboard",
  Snowshoe = "Snowshoe",
  Soccer = "Soccer",
  Squash = "Squash",
  StairStepper = "StairStepper",
  StandUpPaddling = "StandUpPaddling",
  Surfing = "Surfing",
  Swim = "Swim",
  TableTennis = "TableTennis",
  Tennis = "Tennis",
  TrailRun = "TrailRun",
  Velomobile = "Velomobile",
  VirtualRide = "VirtualRide",
  VirtualRow = "VirtualRow",
  VirtualRun = "VirtualRun",
  Walk = "Walk",
  WeightTraining = "WeightTraining",
  Wheelchair = "Wheelchair",
  Windsurf = "Windsurf",
  Workout = "Workout",
  Yoga = "Yoga",
}

/** An enumeration of the types an activity may have. Note that this enumeration does not include new sport types (e.g. MountainBikeRide, EMountainBikeRide), activities with these sport types will have the corresponding activity type (e.g. Ride for MountainBikeRide, EBikeRide for EMountainBikeRide) */
export enum ActivityType {
  AlpineSki = "AlpineSki",
  BackcountrySki = "BackcountrySki",
  Canoeing = "Canoeing",
  Crossfit = "Crossfit",
  EBikeRide = "EBikeRide",
  Elliptical = "Elliptical",
  Golf = "Golf",
  Handcycle = "Handcycle",
  Hike = "Hike",
  IceSkate = "IceSkate",
  InlineSkate = "InlineSkate",
  Kayaking = "Kayaking",
  Kitesurf = "Kitesurf",
  NordicSki = "NordicSki",
  Ride = "Ride",
  RockClimbing = "RockClimbing",
  RollerSki = "RollerSki",
  Rowing = "Rowing",
  Run = "Run",
  Sail = "Sail",
  Skateboard = "Skateboard",
  Snowboard = "Snowboard",
  Snowshoe = "Snowshoe",
  Soccer = "Soccer",
  StairStepper = "StairStepper",
  StandUpPaddling = "StandUpPaddling",
  Surfing = "Surfing",
  Swim = "Swim",
  Velomobile = "Velomobile",
  VirtualRide = "VirtualRide",
  VirtualRun = "VirtualRun",
  Walk = "Walk",
  WeightTraining = "WeightTraining",
  Wheelchair = "Wheelchair",
  Windsurf = "Windsurf",
  Workout = "Workout",
  Yoga = "Yoga",
}
// ...
export interface MetaAthlete {
  /**
   * The unique identifier of the athlete
   * @format int64
   */
  id?: number;
}
// ...
export type LatLng = number[];
// ...
export interface PolylineMap {
  /** The identifier of the map */
  id?: string;
  /** The polyline of the map, only returned on detailed representation of an object */
  polyline?: string;
  /** The summary polyline of the map */
  summary_polyline?: string;
}
// ...
export interface MetaActivity {
  /**
   * The unique identifier of the activity
   * @format int64
   */
  id?: number;
}
// ...
export type SummaryActivity = MetaActivity & {
  /** The identifier provided at upload time */
  external_id?: string;
  /**
   * The identifier of the upload that resulted in this activity
   * @format int64
   */
  upload_id?: number;
  athlete?: MetaAthlete;
  /** The name of the activity */
  name?: string;
  /**
   * The activity's distance, in meters
   * @format float
   */
  distance?: number;
  /** The activity's moving time, in seconds */
  moving_time?: number;
  /** The activity's elapsed time, in seconds */
  elapsed_time?: number;
  /**
   * The activity's total elevation gain.
   * @format float
   */
  total_elevation_gain?: number;
  /**
   * The activity's highest elevation, in meters
   * @format float
   */
  elev_high?: number;
  /**
   * The activity's lowest elevation, in meters
   * @format float
   */
  elev_low?: number;
  /** Deprecated. Prefer to use sport_type */
  type?: ActivityType;
  /** An enumeration of the sport types an activity may have. Distinct from ActivityType in that it has new types (e.g. MountainBikeRide) */
  sport_type?: SportType;
  /**
   * The time at which the activity was started.
   * @format date-time
   */
  start_date?: string;
  /**
   * The time at which the activity was started in the local timezone.
   * @format date-time
   */
  start_date_local?: string;
  /** The timezone of the activity */
  timezone?: string;
  /** A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers. */
  start_latlng?: LatLng;
  /** A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers. */
  end_latlng?: LatLng;
  /** The number of achievements gained during this activity */
  achievement_count?: number;
  /** The number of kudos given for this activity */
  kudos_count?: number;
  /** The number of comments for this activity */
  comment_count?: number;
  /**
   * The number of athletes for taking part in a group activity
   * @min 1
   */
  athlete_count?: number;
  /** The number of Instagram photos for this activity */
  photo_count?: number;
  /** The number of Instagram and Strava photos for this activity */
  total_photo_count?: number;
  map?: PolylineMap;
  /** Whether this activity was recorded on a training machine */
  trainer?: boolean;
  /** Whether this activity is a commute */
  commute?: boolean;
  /** Whether this activity was created manually */
  manual?: boolean;
  /** Whether this activity is private */
  private?: boolean;
  /** Whether this activity is flagged */
  flagged?: boolean;
  /** The activity's workout type */
  workout_type?: number;
  /** The unique identifier of the upload in string format */
  upload_id_str?: string;
  /**
   * The activity's average speed, in meters per second
   * @format float
   */
  average_speed?: number;
  /**
   * The activity's max speed, in meters per second
   * @format float
   */
  max_speed?: number;
  /** Whether the logged-in athlete has kudoed this activity */
  has_kudoed?: boolean;
  /** Whether the activity is muted */
  hide_from_home?: boolean;
  /** The id of the gear for the activity */
  gear_id?: string;
  /**
   * The total work done in kilojoules during this activity. Rides only
   * @format float
   */
  kilojoules?: number;
  /**
   * Average power output in watts during this activity. Rides only
   * @format float
   */
  average_watts?: number;
  /** Whether the watts are from a power meter, false if estimated */
  device_watts?: boolean;
  /** Rides with power meter data only */
  max_watts?: number;
  /** Similar to Normalized Power. Rides with power meter data only */
  weighted_average_watts?: number;
};